Move progress indicator to the generic compose components package
