+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/compose/component/ProgressIndicator.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.ui.login.storecreation
+package com.woocommerce.android.ui.compose.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -11,6 +11,7 @@ import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 
 @Composable
@@ -32,4 +33,10 @@ fun ProgressIndicator(message: String? = null) {
             )
         }
     }
+}
+
+@Preview
+@Composable
+fun ProgressIndicatorPreview() {
+    ProgressIndicator("Loading")
 }
